+++ sources/Task_5.cpp
@@ -4,7 +4,7 @@ bool remove_first_negative_element(const std::vector<int>& vec, int& removed_ele
   std::vector<int> res;
   res = vec;
   removed_element =0;
-  for (int i= 0; i<vec.size();i++){
+  for (int i= 0; i<(int)vec.size();i++){
     if (res[i]<0){
       removed_element = vec[i];
       res.erase(res.begin() +i);
