+++ sources/Task_7.cpp
@@ -8,18 +8,18 @@ std::vector<std::string> split(const std::string& str, char sep){
 
 
 
-  for (int i = 0;i<str.length();i++){
+  for (int i = 0;i<(int) str.length();i++){
     if (str[i] == sep){
       k++;
     }
 
   }
-  if (str.length() == 0 || k == str.length()){
+  if (str.length() == 0 || k == (int) str.length()){
     res = {};
     return res;
   }
 
-  for (int i=0;i<str.length();i++){
+  for (int i=0;i<(int) str.length();i++){
     if (str[i] == sep){
       res.push_back(temp);
       i++;
