+++ sources/Task_8.cpp
@@ -4,7 +4,7 @@ std::string join(const std::vector<std::string>& str, const std::string& sep){
   std::string res;
   res = "";
 
-  for (int i =0;i<str.size()-1;i++){
+  for (int i =0;i<(int)str.size()-1;i++){
     if (str[str.size()] != str[i] ){
       res+=str[i];
       res+=sep;
